Optimize debug output

+++ hook.py
@@ -97,8 +97,9 @@ def _get_txt_record_id(zone_id, name, token):
 
 # https://api.cloudflare.com/#dns-records-for-a-zone-create-dns-record
 def create_txt_record(args):
-    domain, token = args[0], args[2]
+    domain, challenge, token = args
     logger.debug(' + Creating TXT record: {0} => {1}'.format(domain, token))
+    logger.debug(' + Challenge: {0}'.format(challenge))
     zone_id = _get_zone_id(domain)
     name = "{0}.{1}".format('_acme-challenge', domain)
     
@@ -117,7 +118,7 @@ def create_txt_record(args):
     r = requests.post(url, headers=CF_HEADERS, json=payload)
     r.raise_for_status()
     record_id = r.json()['result']['id']
-    logger.debug(" + TXT record created, ID: {0}".format(record_id))
+    logger.debug(" + TXT record created, CFID: {0}".format(record_id))
 
 
 # https://api.cloudflare.com/#dns-records-for-a-zone-delete-dns-record
@@ -131,16 +132,19 @@ def delete_txt_record(args):
     name = "{0}.{1}".format('_acme-challenge', domain)
     record_id = _get_txt_record_id(zone_id, name, token)
 
-    logger.debug(" + Deleting TXT record name: {0}".format(name))
-    url = "https://api.cloudflare.com/client/v4/zones/{0}/dns_records/{1}".format(zone_id, record_id)
-    r = requests.delete(url, headers=CF_HEADERS)
-    r.raise_for_status()
+    if record_id:
+        url = "https://api.cloudflare.com/client/v4/zones/{0}/dns_records/{1}".format(zone_id, record_id)
+        r = requests.delete(url, headers=CF_HEADERS)
+        r.raise_for_status()
+        logger.debug(" + Deleted TXT {0}, CFID {1}".format(name, record_id))
+    else:
+        logger.debug(" + No TXT {0} with token {1}".format(name, token))
 
 
 def deploy_cert(args):
     domain, privkey_pem, cert_pem, fullchain_pem, chain_pem, timestamp = args
-    logger.info(' + ssl_certificate: {0}'.format(fullchain_pem))
-    logger.info(' + ssl_certificate_key: {0}'.format(privkey_pem))
+    logger.debug(' + ssl_certificate: {0}'.format(fullchain_pem))
+    logger.debug(' + ssl_certificate_key: {0}'.format(privkey_pem))
     return
 
 
@@ -149,6 +153,9 @@ def unchanged_cert(args):
     
 
 def invalid_challenge(args):
+    domain, result = args
+    logger.debug(' + invalid_challenge for {0}'.format(domain))
+    logger.debug(' + Full error: {0}'.format(result))
     return
